+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# RcppDynProg 0.1.1 2019/01/19
+
+ * Add logistic scoring.
+
 # RcppDynProg 0.1.0 2019/01/01
 
  * Initial version.